chore: move tests outside src
